webmidi/midiinputmap.html

<!DOCTYPE HTML>
<html>
  <head>
    <meta charset="UTF-8">
    <title>Web MIDI API Test: obtain & list MIDI input ports</title>
    <link rel="author" title="Ryoya KAWAI" href="mailto:ryoya.kawai@gmail.com">
    <link rel="help" href="http://webaudio.github.io/web-midi-api/#midiinputmap-interface">
    <link rel="match" href="reference/REFERENCE-FILE-NAME">
    <meta name="assert" content="Return true when test success in inputs().">
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>

    <style type="text/css">
    /* ADD STYLE BLOCK HERE (PREFERRABLE TO INLINE STYLES) */
    body { font-family:arial;}
    </style>
  </head>
  <body>
    <p>Success when be obtained INPUT MIDI devce and its properties are correct.</p>
    <div id="warning" style="color:red;"></div>

    <div id="log"></div>

    <!-- PAGE CONTENT -->
    <script type="text/javascript">
    var MIDI;
    var timerId={ };
    var getPort={ };

    var test_midiInputMap = async_test("MIDIInputMap Interface Test");

    navigator.requestMIDIAccess({sysex:false}).then( successCallback, errorCallback );
    function successCallback(access) {
        MIDI = access;
        test_midiInputMap.step(function() {
            var inputs=[];
            if(typeof MIDI.inputs==="function") {
                document.getElementById("warning").innerHTML= "This browser seems the implementation of the API, Web MIDI API, is followed old spec.";
                inputs=MIDI.inputs();
            } else {
                var inputIterator=MIDI.inputs.values();
                for(var o=inputIterator.next(); !o.done; o=inputIterator.next()) {
                    inputs.push(o.value);
                }
                console.log("[InputDevices] ", inputs);
                if(inputs.length==0) {
                    document.getElementById("warning").innerHTML="Your Computer looks no <b><u>MIDI INPUT</u></b> devices are pluged-in.<br> Please plug-in at least one <b><u>MIDI INPUT</u></b> device before run this Test.";
                    assert_false(true);
                    test_midiInputMap.done();
                } else {
                    test(function() {
                        assert_true(MIDI.inputs.constructor.name=="MIDIInputMap");
                    }, "MIDIInputMap interface name Test.");
                    test(function() {
                        assert_true(typeof MIDI.inputs.size=="number");
                    }, "MIDIInputMap interface attribute test.");
                    test(function() {
                        assert_true(typeof MIDI.inputs.keys=="function");
                        assert_true(typeof MIDI.inputs.entries=="function");
                        assert_true(typeof MIDI.inputs.values=="function");
                        assert_true(typeof MIDI.inputs.get=="function");
                        assert_true(typeof MIDI.inputs.has=="function");
                    }, "MIDIInputMap interface method test.");
                    assert_true(inputs instanceof Object);
                    for( var j=0; j<inputs.length; j++ ) {
                        test(function() {
                            assert_true( typeof inputs[j]["id"]=="string", "[Type of \"id\": String] " );
                            assert_true( typeof inputs[j]["manufactuerer"]=="string" || inputs[j]["manufactuerer"]==null, "[Type of \"manufactuerer\": String or null] " );
                            assert_true( typeof inputs[j]["name"]=="string" || inputs[j]["name"]==null, "[Type of \"name\": String or null] " );
                            assert_in_array( inputs[j]["type"], ["input", "output"], "[Type of \"type\": input or output] " );
                            assert_true( typeof inputs[j]["version"]=="string" || inputs[j]["version"]==null, "[Type of \"version\": String or null] " );
                            assert_true(inputs[j].constructor.name=="MIDIInput");
                        }, "MIDIPort interface attribute test.");
                        test(function() {
                            assert_true( typeof inputs[j].onmidimessage=="object");
                        }, "MIDIPort interface MIDIInput attribute test.");
                    }
                }
            }
            test_midiInputMap.done();
        });
    }
    function errorCallback(message) {
        test_midiInputMap.step(function() {
            console.log(message);
            assert_true(false);
            test_midiInputMap.done();
        });
    }


    </script>

  </body>
</html>
